Move functions for DarCollectionTable cell data into separate file
